event_rule_gen/requirements.txt
setuptools
pymongo
pytz
numpy
pandas
PyYAML
